cs5/pyproject.toml
[project]
name = "cs5"
version = "0.1.0"
description = "CS5: Agentic Portfolio Intelligence - MCP + LangGraph + CS1-CS4"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # HTTP Client
    "httpx>=0.28.1",
    # MCP SDK
    "mcp[cli]>=1.26.0",
    # Configuration
    "pydantic-settings>=2.13.1",
    # Logging
    "structlog>=24.4.0",
    # LangGraph + LangChain (Lab 10)
    "langgraph>=1.1",
    "langchain-openai>=1.1",
    "langchain-anthropic>=1.3",
    # Streamlit async fix
    "nest-asyncio>=1.6",
    # Observability
    "prometheus-client>=0.21",
    # Dashboard
    "streamlit>=1.40.0",
    "plotly>=6.5.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
bonus = [
    "mem0ai>=0.1",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
